Update more tests
